src/requirements.txt
torch>=1.7.0
torchvision>=0.8.0
transformers>=4.5.0
pyhealth>=1.1.4
numpy>=1.19.0
pandas>=1.1.0
Pillow>=8.0.0
scikit-learn>=0.24.0
matplotlib>=3.3.0
tqdm>=4.50.0
pydicom>=2.2.0
seaborn>=0.13.0